Post/requirements.txt
alembic==1.10.2
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.2
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
grpcio==1.51.3
grpcio-tools==1.51.3
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
prompt-toolkit==3.0.38
protobuf==4.22.1
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.7.1
redis==4.5.1
six==1.16.0
SQLAlchemy==2.0.6
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
